requirements.txt
Flask
gspread
google-auth-oauthlib
google-generativeai
gunicorn
python-dateutil